moved chart functionality to chart.js
